# Add metrics to stresstest and optimize `cadence` usage